requirements.txt
langchain==0.3.7
langchain-google-genai==2.0.2
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3